(selectdb-cloud) Cancel load task when writing local cache failed if there is no free memory buffer #2281 (#2281)
When there is no available free memory buffer, the data will be written to the cache first and then uploaded to S3 when there is an available free memory buffer. However, if an error occurs during the write process to the local cache, continuing to upload the dirty data from the cache to S3 will result in erroneous data(Bad segment).
Considering that local disk write failures are rare, a simple approach is chosen here, which is to treat the import as a failure directly when a local write failure occurs.