(selectdb-cloud) Use pool executor for S3 client instead of default executor to prevent busy loop (#2895)
The implementation of aws sdk's default executor is piece of shit that would cause a lot of CPU consumption when concurrent request reaches.
![image](https://github.com/selectdb/selectdb-core/assets/43750022/85551a46-a50e-49a7-9c7f-8c5e47971788)
Since it's one lockless implementation which takes advantage of atomic bool's cas operation. But actually it's just waste of CPU.
So this pr makes all the S3 operation triggered by S3 client invoked in threadpool to prevent such CPU consumption.