Pick "[Enhancement](group commit) optimize pre allocated calculation (#30893)" (#3131)
For group commit, the data flow is
```
stream pipe -> group commit sink node -> block load queue -> group commit scan node -> tablet channel
|_______________________________|                             |___________________________________|
outside load                             memory queue                     inside load
```

When group commit is set to async mode, it writes to the Write-Ahead Log (WAL). Specifically, once the group commit sink node writes data into the queue, the WAL is immediately persisted to disk. The 'pre allocated' variable tracks the size of all data in memory during the load process. As data is written to the WAL and persisted, 'pre allocated' should be reduced accordingly. Prior to this PR, the reduction of the 'pre allocated' variable was done only after the import was complete. This approach resulted in a portion of the data, already written to the WAL, still being counted in the 'pre allocated' total, leading to inaccuracies in WAL space calculation. This PR addresses the issue by moving the reduction of the 'pre allocated' variable to immediately after the completion of the group commit sink node operation. This PR  helps to avoid the aforementioned inaccuracies to a certain extent.